Keyboard resize fix (#341)
* Move run_one_game_iter() out of WM_PAINT

Use the function that was already there for the message queue and just run dxgi.run_one_game_iter() in a loop. This should be more in line with how this works in SDL/OpenGL.

Fixes bugged (jittery) resize/move with keyboard through the window menu. Although the game pauses when moved/resized or the window menu is open (also like in SDL/OpenGL).

* Fixed game never stopped running after last commit

* Attempt to gracefully close on logout

* Only close if endsession is not canceled